added the ability to take a screenshot and small bug fix

+++ src/a_main/AppInput.h
@@ -16,8 +16,8 @@ struct AppInput
     struct {
         double posX;
         double posY;
-        int deltaX;
-        int deltaY;
+        double deltaX;
+        double deltaY;
         bool capture = false;
     } mouse;
